FastAPI/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
sqlmodel = "^0.0.8"
sqlalchemy = ">=1.4.17,<=1.4.41"
alembic = "^1.10.2"
authlib = "^1.2.0"
fastapi = "^0.94.0"
pytz = "^2022.7.1"
psycopg2-binary = "^2.9.5"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"
pyjwt = "^2.6.0"
python-dateutil = "^2.8.2"
pydantic = {extras = ["dotenv"], version = "^1.10.6"}
fastapi-mail = "^1.2.6"
uvicorn = {extras = ["standard"], version = "^0.21.0"}
password-validator = "^1.0"
jinja2 = "^3.1.2"
sqlalchemy-utils = "^0.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
httpx = "^0.23.3"
requests = "^2.28.2"
types-jinja2 = "^2.11.9"
types-sqlalchemy = "^1.4.53.33"
types-pytz = "^2022.7.1.2"
types-psycopg2 = "^2.9.21.8"
types-passlib = "^1.7.7.10"
types-python-dateutil = "^2.8.19.10"
types-pyjwt = "^1.7.1"
types-requests = "^2.28.11.15"
mypy = "^1.1.1"
pytest-monitor = "^1.6.5"
types-sqlalchemy-utils = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-x --cache-clear --failed-first --durations=0"
